requirements.txt
flask
pandas
numpy
joblib
gunicorn
matplotlib.pyplot
sklearn.model_selection
sklearn.ensemble
sklearn.metrics
sklearn.preprocessing